♻️ separate cmd state & types into their own files
